requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-text-splitters
sentence-transformers
transformers
huggingface-hub
torch
pdfplumber
faiss-cpu
tqdm
